Add a public fbink_close function
For applications which made use of fbink_open to keep fb & mmap around.
This allows us to get rid of the obscure should-have-stayed-internal fb
pointer, as well as weird unbalanced mmap handling (before that, you had
to unmap something you weren't quite sure how/when/if it was evver
mmaped).

No behavior change for applications that always pass -1 as the fbfd
(i.e., letting the library handle this internally on a per function call
basis).