Import Leggie 3.2
First successful BDF conversion ;).

NOTE: It's originally 9x18, I cropped the final column.
(w/ a S&R on the BBOX/Device Width stuff, because I couldn't get it to
stick through gbdfed).
The BDF file imported in the repo is untouched, but the Hex version
was exported from the tweaked one.