Avoid dangling temp realloc pointers
And skip the final shrink + termination, we don't need it for raw data,
we're not buffering C strings.
And we don't care about the potentially extra allocated memory tied up,
since we're going to discard that buffer two lines later.
(And anyway, we'd probably much rather prefer an extra chunk of wasted
memory than the potential copy to another block the shrink might cause
realloc to do).