More accurate comments ;).
NOTE: Caecilia crashed because passing stbtt_GetFontOffsetForIndex's
return value directly to stbtt_InitFont was a terrible idea,
as it returns -1 on error, which obviously leads to terrible results
when used as an offset into the buffer ;).
The Bold variant crashed because it unfortunately holds a large number
at the offset where stbtt expect the number of tables to be stored,
which makes stbtt__find_table loop an inordinate amount of time,
until it obviously starts blowing past the buffer.