Make reinit slightly more aggressive to catch some more corner-cases
Namely, on a fresh boot that includes the processing of KFMon-watched
icons, my Forma is @ 1920x1440, 16bpp @ rotation: 0 during KFMon's init,
which gets flagged isNTX16bLandscape and pickel,
then @ 1440x1920, 16bpp @ rotation: 3 during Nickel's processing of the
file.
Notice that while Nickel has already rotated, it hasn't changed the
bitdepth!
So this no longer gets flagged as isNTX16bLandscape, and as such,
we never pick up on the fact that Nickel switched to 32bpp on the next
reinit, because the canRotate branch only cares about rotation, and that
was already correct.

The good news is: stuff appeared to be rotated properly, but KFMon
prints were then broken in nickel ;).

Fix that by making the 16bpp branch of reinit actually check for that,
16bpp, instead of isNTX16bLandscape.