Jot down some notes about pickel tests on the Forma...
Fun fact: as @frostschutz mentioned, it segfaults without a few env var
properly set \o/. (Prodably PRODUCT & PLATFORM, to be siphoned off
nickel).

In any case: as pickel now prefers a Portrait orientation, that avoids
the whole Landscape mess.
The native boot orientation still appears to be Landscape-y,
and we might incorrectly be naming it "pickel",
but applying the pickel quirks instead of the boot ones
appear to be the right thing to do.

Yes, my brain hurts. :D.

But at least stuff generally behaves properly despite that :D.