Avoid clipping the final glyph of a line.
That's some serious edge-case, usually, serifs with a large RSB and a 0
LSB (f, g, r, especially in italics).
If that glyph's advance positions us a bit early (by design), the final
space ends up making our line width bounding box go backwards, clipping
the edge of the last glyph.
That happens because we don't print a space if it's at the end of a
line, so we don't honor its advance, and we can't correct for it via
LSB, because it's 0 for a space ;).
So we're left with honoring the previou's glyph's gw, which should match
its full bounding box, RSB included.