Speed up put_pixel_RGB32 a tiny bit
Turns out when you bench the right thing, using FBInkPixelBGRA
does bring a small speed bost :D.
Note that the order in which the values are set affects the generated
code a tiny bit.
The order chosen here (which is the same as in fbink_print_image,
incidentally), seems to yield the best code.