Avoid a put_pixel call in the simple cell rendering case
We know we won't ever need the bounds check here, and doing the switch
manually appears to be a tiny bit faster, even if we were to force
inline all the put_pixel* family of functions.

Don't touch the complex cases, as there's also a get_pixel involved,
and we'd have to take care of not fucking up the coordinates rotation
between get & put (which is why the generic put/get takes a value, while
the specific ones take a reference).