More thorough sanity checking of the size_pt float value...
Technically, for input from the CLI tool, iszero() ought to be enough,
as the CLI already sanitizes input.

On the other hand, for API users, this is the safest approach ;).

In any case, since the iszero() macro isn't available in the crappy old
glibc versions we target, that was an easy choice to make ;).