Fix a nasty issue when identifying Kindles...
The swap between old and new device id formats was a bit naive, and
assumed that new S/Ns could never contain a valid old ID...
Turns out, this doesn't hold for Indian PW4s, because they start with
G0010n, and 01 was a valid old ID... for the K1 ;).

Handle the swap by checking the first character instead, which we know
to be limited to B or 9 for the old scheme (and, so far, to G for the
new).