Make get_wfm_mode accurate on the reMarkable
Hopefully. The few constants inherited from the kernel are confusing.
According to the libremarkable comments, I'm fairly confident on the A2
mapping here, but the GL16 one only makes sense by virtue of it usually
always being A2 + 1 on all the platforms I'm familiar with...

TL;DR: Here be dragons!