Make the type coercion we were somewhat silently enforcing on our enums explicit.
Use packed enums to handle the C side of things, and go through an
indirect explicit typedef to make FFI bindings understand the trickery,
since packed enums are a highly non-standard and potentially
unpredictable language extension.
This also makes the actual API/ABI less reliant on said language
breaking trickery ;p.

/!\ ABI break:
fbink_add_ot_font was taking a FONT_STYLE_T argument (which used to be
an int, and is not an uint8_t).