Simplify the read loop handling
This is possibly a remnant of a libevdev-less design, or of testing on a
slower device; but if we're not horrendously slow, stuff should come in
frame by frame, no need to be cute.

(i.e., if you happen to stick a sleep after a SYN_REPORT, then, yeah,
you won't necessarily go back to poll, but that's stupid ;)).