Silence a stringop warning
Both of these arrays are fixed-size, we don't need to try to enforce
truncation and/or a final NUL by twisting strncpy's arm,
just use strcpy.

If the source "string" isn't already NUL terminated, that's a bug in
fbink_device_id, so it's not state's job to try to patch it up.