Clarify that comment a bit
It was written at a time where this didn't use FULL at all,
so fencing it didn't make sense.

It turns out that the stock boot anim *does* display its first iteration
as FULL on color MTK devices, which is why this now does this too,
fencing it off for good measure ;).